Update userReg.html to save user information to the database
This commit modifies the userReg.html file to include code that saves user information to the database upon successful registration. The commented out code for saving user data has been uncommented and the necessary fields such as name, email, phone, address, and isAdmin are being saved. This change improves the functionality of the user registration process and ensures that user data is properly stored in the database.