refactor: Update event attendance button text for closed events
This commit updates the event attendance button text in the events.html file. Previously, when an event was closed, the button displayed "Event is closed". However, it has been changed to "Event is Full!" to provide clearer information to users. This change improves the user experience by accurately reflecting the status of closed events.