refactor: Update admindash.html to display sort dropdown for events
This commit modifies the admindash.html file to add a sort dropdown for events. It adds a new container with the id "sortEvents" and a corresponding select element. Users can now select either "Date" or "Title" to sort the events accordingly. This change enhances the usability of the admin dashboard by providing a convenient way to sort the events based on different criteria.